Supabase and database integrated. Considering regression to Websockets
As of currently:
* Postgres and PGAdmin are running within the `store` directory
* Supabase is running at the `alert` directory
* Supabase is utilizing the `store` database as it's backend
* Nginx is set up to proxy to both services

However, supabase has been a bit of a pain in the ass to set up,
now that it's running, I really like it. However, it does seem like
it may be more overhead than it's worth, especially considering the
auth situation. For that reason, I may try and throw together a MVP
websocket implementation is Flask. If it's something that can be done
in less than ~6 hours, it will be what is used moving foward.