Alert endpoint rewritten in Python
The Alert endpoint has now been rewritten completely in Python, using
the Flask Framework and Flask-SocketIO. The Readme found in the /alert
directory details all the information needed to write a new client.

Important Notes:
* This client does not implement the session creating logic. It was
decided that that logic would be moved to the User API. Instead, it
simply expects a `room` value on the `joined` event which should
contain the session key. This can be returned by the Auth service.